Skip List
This is a data structure for a more efficient search, insert, and
delete. The data structure contains "levels", each with a head node and
each "level" is a linked list of a id number and object, sorted
numerically from smallest to greatest according to the id number. The
amount of levels is dependent of the probability that a new level needs
to be created every time a new id and object is inserted. Please read
the strategy.txt for more in depth description