Multi-level Hash Data Structure
This data structure expands and collapses due to the amount of items it
needs to store in the repository. It uses a Hash function to determine
where to place objects with a certain ID. The data structure is used in
a Service Center program where certain vehicles are stored into it. Each
vehicle has a unique id and the data structure will store the vehicle
according to the hash function. Read the Design.txt for a more in depth
description.